reorg schedule manager and version helper
