+++ src/cppmyth/MythScheduleHelper76.h
@@ -1,0 +1,41 @@
+#pragma once
+/*
+ *      Copyright (C) 2005-2015 Team XBMC
+ *      http://www.xbmc.org
+ *
+ *  This Program is free software; you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation; either version 2, or (at your option)
+ *  any later version.
+ *
+ *  This Program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with XBMC; see the file COPYING.  If not, see
+ *  <http://www.gnu.org/licenses/>.
+ *
+ */
+
+#include "MythScheduleHelper75.h"
+
+// News in 0.27
+
+class MythScheduleHelper76 : public MythScheduleHelper75
+{
+public:
+  MythScheduleHelper76(MythScheduleManager *manager, Myth::Control *control)
+  : MythScheduleHelper75(manager, control) { }
+
+  virtual bool FillTimerEntry(MythTimerEntry& entry, const MythRecordingRuleNode& node) const;
+  virtual MythRecordingRule NewFromTimer(const MythTimerEntry& entry, bool withTemplate);
+
+  // deprecated
+  virtual MythScheduleManager::RuleSummaryInfo GetSummaryInfo(const MythRecordingRule& rule) const;
+  virtual MythRecordingRule NewDailyRecord(const MythEPGInfo& epgInfo);
+  virtual MythRecordingRule NewWeeklyRecord(const MythEPGInfo& epgInfo);
+  virtual MythRecordingRule NewChannelRecord(const MythEPGInfo& epgInfo);
+  virtual MythRecordingRule NewOneRecord(const MythEPGInfo& epgInfo);
+};
